fix: move download progress tests to correct spec file
- Removed incorrectly named .test.ts file
- Added tests to existing ipfs-file-manager.service.spec.ts
- Fixed imports and enum usage
